% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightparser-package.R
\docType{package}
\name{lightparser-package}
\alias{lightparser}
\alias{lightparser-package}
\title{lightparser: From Rmarkdown and Quarto to Tibble and Back}
\description{
Split your Rmarkdown or Quarto file by sections into a tibble: titles, text, chunks. Rebuild it from the tibble.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/lightparser}
  \item \url{https://thinkr-open.github.io/lightparser/}
  \item Report bugs at \url{https://github.com/ThinkR-open/lightparser/issues}
}

}
\author{
\strong{Maintainer}: Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
